# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Hotel Booking API
  version: 0.1.0
  description: |
    Backend HTTP API for the Hotel Booking SaaS. Powers the booking-web
    (guest-facing) and admin-web (operator-facing) Next.js apps.

    All authenticated endpoints expect a JWT access token in the
    `Authorization: Bearer <token>` header. Error responses use a shared
    envelope: `{ "error": { "code": "STRING", "message": "STRING" } }`.

    Money fields ending in `_cents` are integers in the currency's minor unit
    (e.g. THB satang); decimal-string fields (`rate`, `amount`, etc.) are
    used when JS precision matters.
  contact:
    name: Hotel Booking Engineering
    email: engineering@example.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.example.com
    description: Production
...
paths:
...
  /v1/hotels/{hotel_id}/bookings:
    parameters:
      - $ref: '#/components/parameters/HotelIdPathSegment'
    get:
      tags: [bookings]
      summary: List bookings for a hotel (optionally filtered by status)
      parameters:
        - in: query
          name: status
          required: false
          schema: { $ref: '#/components/schemas/BookingStatus' }
        - in: query
          name: limit
          required: false
          schema: { type: integer, minimum: 1, maximum: 500 }
      responses:
        '200':
          description: Booking list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingListResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [bookings]
      summary: Create a walk-in / phone booking (owner / manager / front_desk)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BookingCreateRequest' }
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/hotels/{hotel_id}/bookings/{id}:
    parameters:
      - $ref: '#/components/parameters/HotelIdPathSegment'
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [bookings]
      summary: Get a booking by id (scoped to caller's account)
      responses:
        '200':
          description: Booking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/hotels/{hotel_id}/bookings/{id}/events:
    parameters:
      - $ref: '#/components/parameters/HotelIdPathSegment'
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [bookings]
      summary: List the audit-trail events for a booking (owner / manager / front_desk)
      description: |
        Returns every row of `booking_events` for this booking, ordered by
        `created_at` ascending. Tenant isolation is enforced: a booking in
        another account returns 404 (not 403).
      responses:
        '200':
          description: Event list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingEventsResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/hotels/{hotel_id}/bookings/{id}/confirm:
    parameters:
      - $ref: '#/components/parameters/HotelIdPathSegment'
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [bookings]
      summary: Confirm a pending booking (owner / manager / front_desk)
      responses:
        '200':
          description: Confirmed booking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/hotels/{hotel_id}/bookings/{id}/cancel:
    parameters:
      - $ref: '#/components/parameters/HotelIdPathSegment'
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [bookings]
      summary: Cancel a booking by staff (owner / manager / front_desk)
      description: Body is optional. When provided, `reason` is stored on the booking.
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BookingCancelRequest' }
      responses:
        '200':
          description: Cancelled booking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/hotels/{hotel_id}/bookings/{id}/check-in:
    parameters:
      - $ref: '#/components/parameters/HotelIdPathSegment'
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [bookings]
      summary: Check a guest in (owner / manager / front_desk)
      responses:
        '200':
          description: Checked-in booking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/hotels/{hotel_id}/bookings/{id}/check-out:
    parameters:
      - $ref: '#/components/parameters/HotelIdPathSegment'
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [bookings]
      summary: Check a guest out (owner / manager / front_desk)
      responses:
        '200':
          description: Checked-out booking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/hotels/{hotel_id}/bookings/{id}/no-show:
    parameters:
      - $ref: '#/components/parameters/HotelIdPathSegment'
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [bookings]
      summary: Mark a booking as no-show (owner / manager)
      responses:
        '200':
          description: No-show booking
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ----- subscription -----
...
  /v1/public/bookings/{reference}:
    parameters:
      - $ref: '#/components/parameters/ReferencePath'
    get:
      tags: [public, bookings]
      summary: Look up a booking by reference + guest email
      description: |
        Email is required as a soft auth check. Returns 404 on mismatch.
        Response wraps the `Booking` with hotel context (timezone,
        currency, optional PromptPay receiver) for guest-side rendering.
      security: []
      parameters:
        - in: query
          name: email
          required: true
          schema: { type: string, format: email }
      responses:
        '200':
          description: Booking with hotel context
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicBookingResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/public/bookings/{reference}/cancel:
    parameters:
      - $ref: '#/components/parameters/ReferencePath'
    post:
      tags: [public, bookings]
      summary: Guest-initiated cancel by reference + email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublicBookingCancelRequest' }
      responses:
        '200':
          description: Cancelled booking with hotel context
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicBookingResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/public/hotels/{slug}/bookings:
    parameters:
      - $ref: '#/components/parameters/SlugPath'
    post:
      tags: [public, bookings]
      summary: Create a guest booking by hotel slug
      description: Only hotels with status='live' accept public bookings.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BookingCreateRequest' }
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Booking' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }

# ---------------------------------------------------------------------------
# Components
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued by /v1/auth/login or /v1/auth/refresh.
  parameters:
...
    HotelIdPathSegment:
      in: path
      name: hotel_id
      required: true
      description: Hotel id (parent path segment for sub-resources)
      schema: { type: string, format: uuid }
    IdPath:
      in: path
      name: id
      required: true
      description: Resource id
      schema: { type: string, format: uuid }
...
    SlugPath:
      in: path
      name: slug
      required: true
      description: Public hotel slug
      schema: { type: string, minLength: 3, maxLength: 80, pattern: '^[a-z0-9]([a-z0-9-]*[a-z0-9])?$' }
    ReferencePath:
      in: path
      name: reference
      required: true
      description: Booking reference (e.g. `BK-XXXXXXXX`)
      schema: { type: string }
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid credentials / token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Caller's role does not permit this action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    NotFound:
      description: Resource not found (or hidden by tenant isolation)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Conflict:
      description: State / uniqueness conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine code (e.g. NOT_FOUND, BAD_REQUEST)
            message: { type: string }

    # ----- auth -----
...
    PublicHotelContext:
      type: object
      required: [name, slug, timezone, currency]
      properties:
        name: { type: string }
        slug: { type: string }
        timezone: { type: string, example: 'Asia/Bangkok' }
        currency: { type: string, example: 'THB' }
        promptpay_id:
          type: [string, 'null']
          description: |
            PromptPay receiver (10-digit phone, 13-digit national ID, or
            15-char tax ID). Present only when the hotel has configured one.
...
    BookingStatus:
      type: string
      enum:
        - pending_payment
        - confirmed
        - cancelled
        - expired
        - checked_in
        - checked_out
        - no_show
        - completed
    BookingPaymentStatus:
      type: string
      enum: [pending, paid, refunded, partial, failed]
    BookingSource:
      type: string
      enum: [web, walk_in, phone, admin]
    Booking:
      type: object
      required:
        - id
        - reference
        - hotel_id
        - room_type_id
        - room_count
        - guest_email
        - guest_name
        - check_in_date
        - check_out_date
        - nights
        - currency
        - room_subtotal_cents
        - taxes_cents
        - fees_cents
        - discounts_cents
        - total_cents
        - status
        - payment_status
        - source
        - created_at
        - updated_at
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        hotel_id: { type: string, format: uuid }
        room_type_id: { type: string, format: uuid }
        room_count: { type: integer, minimum: 1 }
        guest_email: { type: string, format: email }
        guest_phone: { type: string }
        guest_name: { type: string }
        guest_country: { type: string }
        special_request: { type: string }
        check_in_date: { type: string, format: date }
        check_out_date: { type: string, format: date }
        nights: { type: integer, minimum: 1 }
        currency: { type: string }
        room_subtotal_cents: { type: integer, format: int64, minimum: 0 }
        taxes_cents: { type: integer, format: int64, minimum: 0 }
        fees_cents: { type: integer, format: int64, minimum: 0 }
        discounts_cents: { type: integer, format: int64, minimum: 0 }
        total_cents: { type: integer, format: int64, minimum: 0 }
        status: { $ref: '#/components/schemas/BookingStatus' }
        payment_status: { $ref: '#/components/schemas/BookingPaymentStatus' }
        payment_method: { type: string }
        expires_at: { type: [string, 'null'], format: date-time }
        cancelled_at: { type: [string, 'null'], format: date-time }
        cancelled_by: { type: string }
        cancellation_reason: { type: string }
        source: { $ref: '#/components/schemas/BookingSource' }
        utm_source: { type: string }
        utm_medium: { type: string }
        utm_campaign: { type: string }
        utm_term: { type: string }
        utm_content: { type: string }
        referrer: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        confirmed_at: { type: [string, 'null'], format: date-time }
        checked_in_at: { type: [string, 'null'], format: date-time }
        checked_out_at: { type: [string, 'null'], format: date-time }
    BookingCreateRequest:
      type: object
      required: [room_type_id, room_count, check_in_date, check_out_date, guest_email, guest_name]
      properties:
        room_type_id: { type: string, format: uuid }
        room_count: { type: integer, minimum: 1 }
        check_in_date: { type: string, format: date }
        check_out_date: { type: string, format: date }
        guest_email: { type: string, format: email }
        guest_phone: { type: string }
        guest_name: { type: string }
        guest_country: { type: string }
        special_request: { type: string }
        currency: { type: string }
        room_subtotal_cents: { type: integer, format: int64, minimum: 0 }
        taxes_cents: { type: integer, format: int64, minimum: 0 }
        fees_cents: { type: integer, format: int64, minimum: 0 }
        discounts_cents: { type: integer, format: int64, minimum: 0 }
        total_cents: { type: integer, format: int64, minimum: 0 }
        utm_source: { type: string }
        utm_medium: { type: string }
        utm_campaign: { type: string }
        utm_term: { type: string }
        utm_content: { type: string }
        referrer: { type: string }
    BookingCancelRequest:
      type: object
      properties:
        reason: { type: string }
    PublicBookingCancelRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        reason: { type: string }
    BookingListResponse:
      type: object
      required: [bookings, total]
      properties:
        bookings:
          type: array
          items: { $ref: '#/components/schemas/Booking' }
        total: { type: integer, minimum: 0 }
    PublicBookingResponse:
      description: |
        Public booking payload. Inherits every field of `Booking` and adds a
        `hotel` block with the timezone + currency + promptpay_id the guest
        UI needs to render dates and a PromptPay QR without a second
        round-trip.
      allOf:
        - $ref: '#/components/schemas/Booking'
        - type: object
          required: [hotel]
          properties:
            hotel: { $ref: '#/components/schemas/PublicHotelContext' }
    BookingEvent:
      type: object
      required: [id, booking_id, event_type, payload, created_at]
      properties:
        id: { type: string, format: uuid }
        booking_id: { type: string, format: uuid }
        event_type:
          type: string
          description: |
            Lifecycle transition or audit marker. Examples: `created`,
            `confirmed`, `cancelled`, `checked_in`, `checked_out`,
            `no_show`.
        actor_type:
          type: string
          description: One of `guest`, `hotel_staff`, `system`, `platform_admin`.
        actor_id: { type: [string, 'null'], format: uuid }
        payload:
          description: Opaque JSON. Shape depends on `event_type`.
        created_at: { type: string, format: date-time }
    BookingEventsResponse:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items: { $ref: '#/components/schemas/BookingEvent' }

    # ----- notification -----
...
security:
  - bearerAuth: []